+++ src/QueryBuilder/QueryBuilder/helpers/valueBuilder.js
@@ -1,6 +1,7 @@
 import { dayjs } from '@folio/stripes/components';
 import { DATA_TYPES } from '../../../constants/dataTypes';
 import { OPERATORS } from '../../../constants/operators';
+import {getOperatorType} from "./selectOptions";
 
 export const getCommaSeparatedStr = (arr) => {
   const str = arr?.map(el => `"${el?.value}"`).join(',');
@@ -68,4 +69,31 @@ export const valueBuilder = ({ value, field, operator, fieldOptions, intl, timez
   };
 
   return valueMap[dataType]?.();
+};
+
+export const retainValueOnOperatorChange = (
+    prevOperator,
+    newOperator,
+    prevValue,
+) => {
+  const prevType = getOperatorType(prevOperator);
+  const newType = getOperatorType(newOperator);
+
+  if (!prevType || !newType) {
+    return ''
+  }
+
+  if (prevType === newType) {
+    return prevValue;
+  }
+
+  if (prevType === 'comparison' && newType === 'comparison array') {
+    return prevValue;
+  }
+
+  if (prevType === 'comparison array' && newType === 'comparison') {
+    return  Array.isArray(prevValue) ? prevValue[0] : prevValue
+  }
+
+  return ''
 };